Inline single-use modifiers next to their call sites
Each of these helpers lived in UI/Primitives/Modifiers (or Controls) but was
used in exactly one place. Following the convention of keeping a type next to
its single use site, move each one into the file that uses it as a private
extension and drop the now-empty shared files:

- alphabetKeyboard() and SearchFieldPlacement.pinned -> AnalyticsView
- pagedTabs() -> OnboardingView
- presentationHeight(_:) -> FilterView
- NSColor.systemGray3 -> MiniChart, NSColor.systemGray5 -> ReleaseRow
- SeeAllButton -> EventView.Sections

Header's preview no longer depends on SeeAllButton, so it uses an inline
button instead. Doc comments demoted with the helpers become brief // notes,
per the public-API-only documentation rule.
