+++ Sources/Scout/UI/Chart/View/MiniChart.swift
@@ -57,6 +57,21 @@ struct MiniChart: View {
     }
 }
 
+#if os(macOS)
+    import AppKit
+
+    private extension NSColor {
+        // Stand-in for the iOS-only tiered system gray, matching its light and dark values.
+        static var systemGray3: NSColor {
+            NSColor(name: nil) { appearance in
+                appearance.bestMatch(from: [.darkAqua, .aqua]) == .darkAqua
+                    ? NSColor(red: 72 / 255, green: 72 / 255, blue: 74 / 255, alpha: 1)
+                    : NSColor(red: 199 / 255, green: 199 / 255, blue: 204 / 255, alpha: 1)
+            }
+        }
+    }
+#endif
+
 #Preview {
     NavigationStack {
         List {
