+++ Sources/Scout/UI/Event/Detail/EventView.Sections.swift
@@ -110,3 +110,14 @@ extension EventView {
         }
     }
 }
+
+private struct SeeAllButton: View {
+    let action: () -> Void
+
+    var body: some View {
+        Button(action: action) {
+            Text(verbatim: "See all").foregroundStyle(.blue)
+        }
+        .buttonStyle(.plain)
+    }
+}
