--- Sources/Scout/UI/Primitives/Modifiers/SearchFieldPlacement+Pinned.swift
@@ -1,19 +1,0 @@
-//
-// Copyright 2026 Mikhail Kasianov
-//
-// Use of this source code is governed by an MIT-style
-// license that can be found in the LICENSE file or at
-// https://opensource.org/licenses/MIT.
-
-import SwiftUI
-
-extension SearchFieldPlacement {
-    /// An always-visible search field: pinned under the navigation bar on iOS, automatic on macOS.
-    static var pinned: SearchFieldPlacement {
-        #if os(iOS)
-            .navigationBarDrawer(displayMode: .always)
-        #else
-            .automatic
-        #endif
-    }
-}
