+++ Sources/Scout/UI/Releases/View/ReleaseRow.swift
@@ -82,6 +82,21 @@ private struct ReleasePercent: View {
     }
 }
 
+#if os(macOS)
+    import AppKit
+
+    private extension NSColor {
+        // Stand-in for the iOS-only tiered system gray, matching its light and dark values.
+        static var systemGray5: NSColor {
+            NSColor(name: nil) { appearance in
+                appearance.bestMatch(from: [.darkAqua, .aqua]) == .darkAqua
+                    ? NSColor(red: 44 / 255, green: 44 / 255, blue: 46 / 255, alpha: 1)
+                    : NSColor(red: 229 / 255, green: 229 / 255, blue: 234 / 255, alpha: 1)
+            }
+        }
+    }
+#endif
+
 #Preview {
     NavigationStack {
         List {
